refactor(eventbus): split rpc.rs monolith into rpc/ submodules
rpc.rs was 2789 lines — ~4.5x the project's 500-600 line target (AGENTS.md
rule 09). It contained four distinct responsibilities with clean seams, now
extracted into a module directory:

- rpc/messages.rs  (687) — request/response & data types + their builders
- rpc/providers.rs (413) — Health/Config/Registration provider traits + defaults
- rpc/client.rs    (231) — CollectorRpcClient
- rpc/service.rs  (1336) — CollectorRpcService (one cohesive impl, left intact)
- rpc/tests.rs     (118) — the inline unit tests, verbatim
- rpc/mod.rs        (59) — module doc + re-exports preserving the public API

Pure mechanical extraction: zero logic/signature/API changes. Private items
kept private via pub(super) (matching the original module-private semantics)
so nothing leaks into the public surface. Verified: build, 88 lib tests, 10
doctests, and clippy -D warnings all green.

Also fixes a pre-existing broken doctest in lib.rs (missing include_control
field on EventSubscription) surfaced while verifying.
